Simplify task view tests
